# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Voxus Location",
        "description": "Documentação do Projeto Voxus Location",
        "contact": {
            "name": "João Paulo Pinheiro",
            "email": "joaopinheiro.ti@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://voxus.local/api/v1",
            "description": "Ambiente de Desenvolvimento"
        },
        {
            "url": "http://voxus.prod/api/v1",
            "description": "Ambiente de Produção"
        }
    ],
    "paths": {
...
        "/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Listar as Localizações por ID de Usuário.",
                "operationId": "App\\Http\\Controllers\\LocationController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/LocationProperty"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Deletar Localização por ID.",
                "operationId": "App\\Http\\Controllers\\LocationController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Localização",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Lista de permissões.",
                "operationId": "App\\Http\\Controllers\\PermissionController::index",
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PermissionsPaginateResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Criar nova permissão.",
                "operationId": "App\\Http\\Controllers\\PermissionController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "display_name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome da Permissão",
                                        "type": "string"
                                    },
                                    "display_name": {
                                        "description": "Nome de Exibição",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Listar permissão por ID.",
                "operationId": "App\\Http\\Controllers\\PermissionController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PermissionProperty"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Atualizar permissão.",
                "operationId": "App\\Http\\Controllers\\PermissionController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "display_name",
                                    "description"
                                ],
                                "properties": {
                                    "display_name": {
                                        "description": "Nome de Exibição",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Deletar permissão.",
                "operationId": "App\\Http\\Controllers\\PermissionController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Lista de perfis.",
                "operationId": "App\\Http\\Controllers\\RoleController::index",
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/RolesPaginateResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Criar novo perfil.",
                "operationId": "App\\Http\\Controllers\\RoleController::store",
                "parameters": [
                    {
                        "name": "permissions[]",
                        "in": "query",
                        "description": "Permissões do Perfil",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "display_name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do Perfil",
                                        "type": "string"
                                    },
                                    "display_name": {
                                        "description": "Nome de Exibição",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Retorno padrão",
                "description": "Entidade de resposta, resultado da resposta usa essa estrutura uniformemente",
                "properties": {
                    "success": {
                        "description": "Status da resposta",
                        "type": "boolean"
                    },
                    "code": {
                        "description": "Código da resposta",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Mensagem da resposta",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LocationProperty": {
                "title": "Localizações",
                "description": "Localizações",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer"
                    },
                    "ip": {
                        "description": "ip",
                        "type": "string"
                    },
                    "latitude": {
                        "description": "latitude",
                        "type": "string"
                    },
                    "longitude": {
                        "description": "longitude",
                        "type": "string"
                    },
                    "cidade": {
                        "description": "cidade",
                        "type": "string"
                    },
                    "estado": {
                        "description": "estado",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserProperty"
                        }
                    }
                },
                "type": "object"
            },
...
            "PermissionProperty": {
                "title": "Auth Permissions",
                "description": "Permissões de Usuário",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "display name",
                        "type": "string"
                    },
                    "description": {
                        "description": "description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PermissionsPaginateResponse": {
                "title": "Permissions Paginate",
                "description": "Lista de Permissões com Paginação",
                "properties": {
                    "current_page": {
                        "description": "current_page",
                        "type": "integer"
                    },
                    "first_page_url": {
                        "description": "first_page_url",
                        "type": "string"
                    },
                    "from": {
                        "description": "from",
                        "type": "integer"
                    },
                    "last_page": {
                        "description": "last_page",
                        "type": "integer"
                    },
                    "last_page_url": {
                        "description": "last_page_url",
                        "type": "string"
                    },
                    "next_page_url": {
                        "description": "next_page_url",
                        "type": "string"
                    },
                    "path": {
                        "description": "path",
                        "type": "string"
                    },
                    "per_page": {
                        "description": "per_page",
                        "type": "integer"
                    },
                    "prev_page_url": {
                        "description": "prev_page_url",
                        "type": "string"
                    },
                    "to": {
                        "description": "to",
                        "type": "integer"
                    },
                    "total": {
                        "description": "total",
                        "type": "integer"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionProperty"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Informe o token JWT para endpoits seguros",
                "name": "JWT bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}